refactor(allocator/pool): move `AllocatorPool` into own directory (#13619)
Pure refactor. Move all code related to `AllocatorPool` into a directory together, in preparation for further changes.
