+++ tasks/ast_tools/src/main.rs
@@ -207,7 +207,7 @@ use utils::create_ident;
 
 /// Paths to source files containing AST types
 static SOURCE_PATHS: &[&str] = &[
-    "crates/oxc_allocator/src/pool_fixed_size.rs",
+    "crates/oxc_allocator/src/pool/fixed_size.rs",
     "crates/oxc_ast/src/ast/js.rs",
     "crates/oxc_ast/src/ast/literal.rs",
     "crates/oxc_ast/src/ast/jsx.rs",
